requirements.txt
fastapi
uvicorn

torch
torchvision

pillow
numpy
opencv-python

scikit-image
scikit-learn
lime

jinja2
python-multipart
